Grant's First Checkpoint Parts
Added my parts that are due Monday.

The Candidate block is the Is There More Transition Block where you should put your blocks in.

The setup block marks the Current State with C and sets it to 1 then goes to the right and marks the stack with Z placing 1 to represent Z.

The Cur2Stk block checks the transitions for a transition from this state. If it finds them, it sets a flag in the Clean Up block. If there is no transition from this state, then it stops at the F of the input string.